Renamed Allocator to RefCountPolicy
