+++ uti.hpp
@@ -12,7 +12,7 @@
 
 #include "uti/utiCommonHeader.hpp"
 #include "uti/utiAllocator.hpp"
-#include "uti/utiAllocatorPolicy.hpp"
+#include "uti/utiRefCountPolicy.hpp"
 #include "uti/utiReferenceCounted.hpp"
 #include "uti/utiByteIterator.hpp"
 #include "uti/utiCharIterator.hpp"
